requirements.txt
rembg==2.0.61
flask==3.1.2
pillow==11.3.0
pywebview==6.1
pyinstaller==6.17.0
onnxruntime==1.19.2
numpy==2.0.2
scipy==1.13.1
scikit-image==0.24.0
opencv-python-headless==4.12.0.88
